credit-risk-analysis/backend/requirements.txt
# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.20

# Database
sqlalchemy>=2.0.36
psycopg[binary]>=3.2.0
alembic>=1.14.0

# Configuration
pydantic>=2.10.0
pydantic-settings>=2.7.0
python-dotenv>=1.0.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.1

# Machine Learning






# Utilities
httpx>=0.28.0

email-validator
